Add `futures_03` feature
refactor `futures-core` as optional
refactor std futures support into `futures_03` module
add `futures_03` as default feature (via `sink` feature)
update documentation
remove `#[cfg(feature="sink")]` on futures_01, as it is not optional
